+++ eduadmin.php
@@ -12,7 +12,7 @@ if(!session_id())
  * Plugin URI:	http://www.eduadmin.se
  * Description:	EduAdmin plugin to allow visitors to book courses at your website
  * Tags:	booking, participants, courses, events, eduadmin, lega online
- * Version:	0.9.11
+ * Version:	0.9.12
  * Requires at least: 3.0
  * Tested up to: 4.7
  * Author:	Chris Gårdenberg, MultiNet Interactive AB
@@ -79,7 +79,7 @@ function edu_call_home()
         'wpVersion' => $wp_version,
         'token' => get_option('eduadmin-api-key'),
         'officialVersion' => file_exists(dirname(__FILE__) . "/.official.plugin.php"),
-        'pluginVersion' => '0.9.11'
+        'pluginVersion' => '0.9.12'
     );
 
     $callHomeUrl = 'http://ws10.multinet.se/edu-plugin/wp_phone_home.php';