+++ includes/_options.php
@@ -40,6 +40,7 @@ function eduadmin_settings_init()
 	register_setting('eduadmin-booking', 'eduadmin-allowInterestRegObject');
 	register_setting('eduadmin-booking', 'eduadmin-allowInterestRegEvent');
 	register_setting('eduadmin-booking', 'eduadmin-hideSubEventDateTime');
+	register_setting('eduadmin-booking', 'eduadmin-allowDiscountCode');
 
 	/* Phrase settings */
 	register_setting('eduadmin-phrases', 'eduadmin-phrases');